+++ core/src/main/java/bisq/core/trade/protocol/bisq_v1/protocol_v4/BaseSellerProtocol_v4.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.core.trade.protocol.bisq_v1;
+package bisq.core.trade.protocol.bisq_v1.protocol_v4;
 
 import bisq.core.trade.model.bisq_v1.SellerTrade;
 import bisq.core.trade.model.bisq_v1.Trade;
@@ -23,6 +23,7 @@ import bisq.core.trade.protocol.FluentProtocol;
 import bisq.core.trade.protocol.SellerProtocol;
 import bisq.core.trade.protocol.TradeMessage;
 import bisq.core.trade.protocol.TradeTaskRunner;
+import bisq.core.trade.protocol.bisq_v1.DisputeProtocol;
 import bisq.core.trade.protocol.bisq_v1.messages.CounterCurrencyTransferStartedMessage;
 import bisq.core.trade.protocol.bisq_v1.messages.DelayedPayoutTxSignatureResponse;
 import bisq.core.trade.protocol.bisq_v1.messages.ShareBuyerPaymentAccountMessage;
@@ -48,13 +49,13 @@ import bisq.common.handlers.ResultHandler;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
-abstract class BaseSellerProtocol extends DisputeProtocol implements SellerProtocol {
+abstract class BaseSellerProtocol_v4 extends DisputeProtocol implements SellerProtocol {
     enum SellerEvent implements FluentProtocol.Event {
         STARTUP,
         PAYMENT_RECEIVED
     }
 
-    protected BaseSellerProtocol(SellerTrade trade) {
+    protected BaseSellerProtocol_v4(SellerTrade trade) {
         super(trade);
     }
